requirements.txt
# Core Langchain
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
pydantic>=2.6.4,<3.0.0
python-dotenv>=1.0.0
typing-extensions>=4.10.0

# Langchain Components
langchain-google-genai>=2.0.0,<3.0.0
langchain-mongodb>=0.1.0

# Database & Vector Store
pymongo>=4.6.0
qdrant-client>=1.8.0
google-cloud-bigquery>=3.19.0

# Google AI
google-generativeai>=0.5.0
google-cloud-aiplatform>=1.48.0

# Utilities
tabulate>=0.9.0